+++ external/xtorch_harmonics/xtorch_harmonics/__init__.py
@@ -1,1 +1,1 @@
-from .core import roundtrip
+from .xtorch_harmonics import roundtrip
